Refactor Typeahead to standard Portal/floating-ui
