+++ packages/grafana-ui/src/components/Typeahead/Typeahead.tsx
@@ -1,14 +1,17 @@
 import { css } from '@emotion/css';
+import { autoUpdate, offset, useFloating } from '@floating-ui/react';
 import { isEqual } from 'lodash';
-import { createRef, PureComponent } from 'react';
+import { createRef, PureComponent, useEffect, type PropsWithChildren } from 'react';
 import * as React from 'react';
-import ReactDOM from 'react-dom';
 import { FixedSizeList } from 'react-window';
 
 import { type GrafanaTheme2, ThemeContext } from '@grafana/data';
 
 import { type CompletionItem, type CompletionItemGroup, CompletionItemKind } from '../../types/completion';
+import { SelectionReference } from '../../utils/SelectionReference';
+import { getPositioningMiddleware } from '../../utils/floating';
 import { flattenGroupItems, calculateLongestLabel, calculateListSizes } from '../../utils/typeahead';
+import { Portal } from '../Portal/Portal';
 
 import { TypeaheadInfo } from './TypeaheadInfo';
 import { TypeaheadItem } from './TypeaheadItem';
@@ -52,8 +55,6 @@ export class Typeahead extends PureComponent<Props, State> {
       this.props.menuRef(this);
     }
 
-    document.addEventListener('selectionchange', this.handleSelectionChange);
-
     const allItems = flattenGroupItems(this.props.groupedItems);
     const longestLabel = calculateLongestLabel(allItems);
     const { listWidth, listHeight, itemHeight } = calculateListSizes(this.context, allItems, longestLabel);
@@ -63,14 +64,6 @@ export class Typeahead extends PureComponent<Props, State> {
       itemHeight,
       allItems,
     });
-  };
-
-  componentWillUnmount = () => {
-    document.removeEventListener('selectionchange', this.handleSelectionChange);
-  };
-
-  handleSelectionChange = () => {
-    this.forceUpdate();
   };
 
   componentDidUpdate = (prevProps: Readonly<Props>, prevState: Readonly<State>) => {
@@ -132,32 +125,8 @@ export class Typeahead extends PureComponent<Props, State> {
     }
   };
 
-  get menuPosition(): string {
-    // Exit for unit tests
-    if (!window.getSelection) {
-      return '';
-    }
-
-    const selection = window.getSelection();
-    const node = selection && selection.anchorNode;
-
-    // Align menu overlay to editor node
-    if (node && node.parentElement) {
-      // Read from DOM
-      const rect = node.parentElement.getBoundingClientRect();
-      const scrollX = window.scrollX;
-      const scrollY = window.scrollY;
-
-      return `position: absolute; display: flex; top: ${rect.top + scrollY + rect.height + 6}px; left: ${
-        rect.left + scrollX - 2
-      }px`;
-    }
-
-    return '';
-  }
-
   render() {
-    const { prefix, isOpen = false, origin } = this.props;
+    const { prefix, isOpen = false } = this.props;
     const { allItems, listWidth, listHeight, itemHeight, hoveredItem, typeaheadIndex } = this.state;
     const styles = getStyles(this.context);
 
@@ -165,7 +134,7 @@ export class Typeahead extends PureComponent<Props, State> {
     const documentationItem = allItems[hoveredItem ? hoveredItem : typeaheadIndex || 0];
 
     return (
-      <Portal origin={origin} isOpen={isOpen} style={this.menuPosition}>
+      <TypeaheadPortal isOpen={isOpen}>
         <ul role="menu" className={styles.typeahead} data-testid="typeahead">
           <FixedSizeList
             ref={this.listRef}
@@ -201,48 +170,55 @@ export class Typeahead extends PureComponent<Props, State> {
         </ul>
 
         {showDocumentation && <TypeaheadInfo height={listHeight} item={documentationItem} />}
-      </Portal>
+      </TypeaheadPortal>
     );
   }
 }
 
-interface PortalProps {
-  index?: number;
-  isOpen: boolean;
-  origin: string;
-  style: string;
-}
+/**
+ * Positions the typeahead menu next to the caret in the Slate editor and renders it into the
+ * shared overlay portal. Uses floating-ui with a selection-anchored virtual reference, the same
+ * pattern used by other caret-anchored menus (e.g. DataLinkInput).
+ */
+function TypeaheadPortal({ isOpen, children }: PropsWithChildren<{ isOpen: boolean }>) {
+  const { refs, floatingStyles, update } = useFloating({
+    open: isOpen,
+    placement: 'bottom-start',
+    strategy: 'fixed',
+    // Mirror the previous manual offsets: 6px below the caret line, nudged 2px to the left.
+    middleware: [offset({ mainAxis: 6, crossAxis: -2 }), ...getPositioningMiddleware('bottom-start')],
+    whileElementsMounted: autoUpdate,
+  });
 
-class Portal extends PureComponent<React.PropsWithChildren<PortalProps>, {}> {
-  node: HTMLElement;
+  // Anchor positioning to the current text selection (the caret). Set once; the virtual element
+  // reads the live selection rect on every reposition.
+  useEffect(() => {
+    refs.setReference(new SelectionReference());
+  }, [refs]);
 
-  constructor(props: React.PropsWithChildren<PortalProps>) {
-    super(props);
-    const { index = 0, origin = 'query' } = props;
-    this.node = document.createElement('div');
-    this.node.classList.add(`slate-typeahead-${origin}-${index}`);
-  }
-
-  componentDidMount() {
-    this.node.setAttribute('style', this.props.style);
-    document.body.appendChild(this.node);
-  }
-
-  componentWillUnmount() {
-    this.node.remove();
-  }
-
-  render() {
-    if (this.props.isOpen) {
-      this.node.setAttribute('style', this.props.style);
-      this.node.classList.add(`slate-typeahead--open`);
-      return ReactDOM.createPortal(this.props.children, this.node);
-    } else {
-      this.node.classList.remove(`slate-typeahead--open`);
+  // The caret moves as the user types or navigates, so reposition on selection changes while open.
+  useEffect(() => {
+    if (!isOpen) {
+      return;
     }
+    document.addEventListener('selectionchange', update);
+    return () => document.removeEventListener('selectionchange', update);
+  }, [isOpen, update]);
 
+  if (!isOpen) {
     return null;
   }
+
+  // The `slate-typeahead--open` class is a behavioural hook: keybindingSrv queries for it so the
+  // global Esc handler defers to the typeahead while suggestions are open.
+  return (
+    <Portal>
+      {/* display: flex lays the suggestion list and the documentation panel out side by side. */}
+      <div ref={refs.setFloating} style={{ ...floatingStyles, display: 'flex' }} className="slate-typeahead--open">
+        {children}
+      </div>
+    </Portal>
+  );
 }
 
 const getStyles = (theme: GrafanaTheme2) => ({
